requirements.txt
discord.py==1.7.3
APScheduler==3.9.1
asyncpg==0.26.0